RPA_LAB/requirement.txt
asgiref==3.4.1
Django==3.2.14
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
